requirements.txt
youtube_transcript_api>=0.6.0
flask>=2.0.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pathlib>=1.0.1
PyPDF2>=3.0.0
python-pptx>=0.6.21
requests>=2.28.0
duckduckgo_search>=3.8.0
werkzeug>=2.0.0